\name{uglypaRse-package}
\alias{uglypaRse-package}
\alias{uglypaRse}
\docType{package}
\title{
A package to split sets of words that are missing delimiters. 
}
\description{
uglypaRse contains tools for dealing with collapsed strings of words, such as 'helloworld' or 'coolshirthadley'. It is currently in development, and should be treated as an alpha release.
}
\details{
}
\author{
Joel Levin <joelmlevin@gmail.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ natural language text word break delimiter }
\seealso{
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
